Create plugin for API
The API implementation has been moved into a custom Bevy plugin to
improve the code organization.
