+++ src/lunaria/src/app.rs
@@ -1,5 +1,7 @@
 use bevy::prelude::*;
+use tokio::sync::broadcast;
 
+use crate::plugin::api::Api;
 use crate::scene::main_menu::MainMenu;
 
 #[derive(Debug, Clone, Eq, PartialEq, Hash)]
@@ -9,6 +11,8 @@ pub enum AppState {
 }
 
 pub fn run_app() {
+    let (command_tx, _command_rx) = broadcast::channel(256);
+
     App::build()
         // Built-in plugins
         .add_plugins(DefaultPlugins)
@@ -18,5 +22,7 @@ pub fn run_app() {
         .insert_resource(ClearColor(Color::hsl(231.0, 0.15, 0.18)))
         // Scenes
         .add_plugin(MainMenu)
+        // Custom plugins
+        .add_plugin(Api::new(command_tx))
         .run();
 }
