+++ src/lunaria/src/lib.rs
@@ -1,7 +1,6 @@
-pub mod api;
 pub mod app;
 pub mod command;
-pub mod engine;
 pub mod error;
 pub mod event;
+pub mod plugin;
 pub mod scene;
